Transformed old sign up page into page in which you select your role before being redirected to that role's signup page. Put single header on each role signup page to ensure correct routing upon clicking each role. Created new file called old_page.tsx to store code for the old signup page for my reference when creating the new sign up pages.
