+++ frontend/src/app/signup/business/page.tsx
@@ -1,0 +1,7 @@
+// Sign up page for business accounts
+
+export default function BusinessSignup() {
+    return (
+        <header>Business Sign Up</header>
+    );
+}
