+++ frontend/src/app/signup/musician/page.tsx
@@ -1,0 +1,7 @@
+// Sign up page for musician accounts
+
+export default function MusicianSignup() {
+    return (
+        <header>Musician Sign Up</header>
+    );
+}
